refactor: split rendering function into multiple files
