+++ src/chart_types/xy_chart/rendering/bubble.ts
@@ -1,0 +1,80 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { Scale } from '../../../scales';
+import { Color } from '../../../utils/commons';
+import { Dimensions } from '../../../utils/dimensions';
+import { BubbleGeometry } from '../../../utils/geometry';
+import { BubbleSeriesStyle } from '../../../utils/themes/theme';
+import { IndexedGeometryMap } from '../utils/indexed_geometry_map';
+import { DataSeries } from '../utils/series';
+import { PointStyleAccessor } from '../utils/specs';
+import { renderPoints } from './points';
+import { MarkSizeOptions } from './utils';
+
+/** @internal */
+export function renderBubble(
+  shift: number,
+  dataSeries: DataSeries,
+  xScale: Scale,
+  yScale: Scale,
+  color: Color,
+  panel: Dimensions,
+  hasY0Accessors: boolean,
+  xScaleOffset: number,
+  seriesStyle: BubbleSeriesStyle,
+  markSizeOptions: MarkSizeOptions,
+  isMixedChart: boolean,
+  pointStyleAccessor?: PointStyleAccessor,
+): {
+  bubbleGeometry: BubbleGeometry;
+  indexedGeometryMap: IndexedGeometryMap;
+} {
+  const { pointGeometries, indexedGeometryMap } = renderPoints(
+    shift - xScaleOffset,
+    dataSeries,
+    xScale,
+    yScale,
+    panel,
+    color,
+    seriesStyle.point,
+    hasY0Accessors,
+    markSizeOptions,
+    pointStyleAccessor,
+    !isMixedChart,
+  );
+
+  const bubbleGeometry = {
+    points: pointGeometries,
+    color,
+    seriesIdentifier: {
+      key: dataSeries.key,
+      specId: dataSeries.specId,
+      yAccessor: dataSeries.yAccessor,
+      splitAccessors: dataSeries.splitAccessors,
+      seriesKeys: dataSeries.seriesKeys,
+      smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
+      smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
+    },
+    seriesPointStyle: seriesStyle.point,
+  };
+  return {
+    bubbleGeometry,
+    indexedGeometryMap,
+  };
+}
