--- src/chart_types/xy_chart/rendering/rendering.ts
@@ -1,919 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { area, line } from 'd3-shape';
-
-import { LegendItem } from '../../../commons/legend';
-import { Scale } from '../../../scales';
-import { ScaleType } from '../../../scales/constants';
-import { isLogarithmicScale } from '../../../scales/types';
-import { MarkBuffer, StackMode } from '../../../specs';
-import { CanvasTextBBoxCalculator } from '../../../utils/bbox/canvas_text_bbox_calculator';
-import { mergePartial, Color, getDistance } from '../../../utils/commons';
-import { CurveType, getCurveFactory } from '../../../utils/curves';
-import { Dimensions } from '../../../utils/dimensions';
-import {
-  PointGeometry,
-  BarGeometry,
-  AreaGeometry,
-  LineGeometry,
-  isPointGeometry,
-  ClippedRanges,
-  BandedAccessorType,
-  BubbleGeometry,
-} from '../../../utils/geometry';
-import {
-  AreaSeriesStyle,
-  LineSeriesStyle,
-  PointStyle,
-  SharedGeometryStateStyle,
-  BarSeriesStyle,
-  GeometryStateStyle,
-  LineStyle,
-  BubbleSeriesStyle,
-  DisplayValueStyle,
-} from '../../../utils/themes/theme';
-import { IndexedGeometryMap, GeometryType } from '../utils/indexed_geometry_map';
-import { DataSeriesDatum, DataSeries, XYChartSeriesIdentifier } from '../utils/series';
-import { DisplayValueSpec, PointStyleAccessor, BarStyleAccessor } from '../utils/specs';
-import { DEFAULT_HIGHLIGHT_PADDING } from './constants';
-
-export interface MarkSizeOptions {
-  enabled: boolean;
-  ratio?: number;
-}
-/**
- * Returns a safe scaling factor for label text for fixed or range size inputs
- * @internal
- */
-function getFinalFontScalingFactor(
-  scale: number,
-  fixedFontSize: number,
-  limits: DisplayValueStyle['fontSize'],
-): number {
-  if (typeof limits === 'number') {
-    // it's a fixed size, so it's always ok
-    return 1;
-  }
-  const finalFontSize = scale * fixedFontSize;
-  if (finalFontSize > limits.max) {
-    return limits.max / fixedFontSize;
-  }
-  if (finalFontSize < limits.min) {
-    // it's technically 1, but keep it generic in case the fixedFontSize changes
-    return limits.min / fixedFontSize;
-  }
-  return scale;
-}
-
-/**
- * Workout the text box size and fixedFontSize based on a collection of options
- * @internal
- */
-function computeBoxWidth(
-  text: string,
-  {
-    padding,
-    fontSize,
-    fontFamily,
-    bboxCalculator,
-    width,
-  }: {
-    padding: number;
-    fontSize: number | { min: number; max: number };
-    fontFamily: string;
-    bboxCalculator: CanvasTextBBoxCalculator;
-    width: number;
-  },
-  displayValueSettings: DisplayValueSpec | undefined,
-): { fixedFontScale: number; displayValueWidth: number } {
-  const fixedFontScale = Math.max(typeof fontSize === 'number' ? fontSize : fontSize.min, 1);
-
-  const computedDisplayValueWidth = bboxCalculator.compute(text || '', padding, fixedFontScale, fontFamily).width;
-  if (typeof fontSize !== 'number') {
-    return {
-      fixedFontScale,
-      displayValueWidth: computedDisplayValueWidth,
-    };
-  }
-  return {
-    fixedFontScale,
-    displayValueWidth:
-      displayValueSettings && displayValueSettings.isValueContainedInElement ? width : computedDisplayValueWidth,
-  };
-}
-
-/**
- * Returns value of `y1` or `filled.y1` or null
- * @internal
- */
-export const getYValue = ({ y1, filled }: DataSeriesDatum): number | null => {
-  if (y1 !== null) {
-    return y1;
-  }
-
-  if (filled && filled.y1 !== undefined) {
-    return filled.y1;
-  }
-
-  return null;
-};
-
-/** @internal */
-export function getPointStyleOverrides(
-  datum: DataSeriesDatum,
-  seriesIdentifier: XYChartSeriesIdentifier,
-  pointStyleAccessor?: PointStyleAccessor,
-): Partial<PointStyle> | undefined {
-  const styleOverride = pointStyleAccessor && pointStyleAccessor(datum, seriesIdentifier);
-
-  if (!styleOverride) {
-    return;
-  }
-
-  if (typeof styleOverride === 'string') {
-    return {
-      stroke: styleOverride,
-    };
-  }
-
-  return styleOverride;
-}
-
-/** @internal */
-export function getBarStyleOverrides(
-  datum: DataSeriesDatum,
-  seriesIdentifier: XYChartSeriesIdentifier,
-  seriesStyle: BarSeriesStyle,
-  styleAccessor?: BarStyleAccessor,
-): BarSeriesStyle {
-  const styleOverride = styleAccessor && styleAccessor(datum, seriesIdentifier);
-
-  if (!styleOverride) {
-    return seriesStyle;
-  }
-
-  if (typeof styleOverride === 'string') {
-    return {
-      ...seriesStyle,
-      rect: {
-        ...seriesStyle.rect,
-        fill: styleOverride,
-      },
-    };
-  }
-
-  return mergePartial(seriesStyle, styleOverride, {
-    mergeOptionalPartialValues: true,
-  });
-}
-
-type GetRadiusFnReturn = (mark: number | null, defaultRadius?: number) => number;
-
-/**
- * Get radius function form ratio and min/max mark size
- *
- * @todo add continuous/non-stepped function
- *
- * @param  {DataSeriesDatum[]} data
- * @param  {number} lineWidth
- * @param  {number=50} markSizeRatio - 0 to 100
- * @internal
- */
-export function getRadiusFn(data: DataSeriesDatum[], lineWidth: number, markSizeRatio: number = 50): GetRadiusFnReturn {
-  if (data.length === 0) {
-    return () => 0;
-  }
-  const { min, max } = data.reduce(
-    (acc, { mark }) =>
-      mark === null
-        ? acc
-        : {
-            min: Math.min(acc.min, mark / 2),
-            max: Math.max(acc.max, mark / 2),
-          },
-    { min: Infinity, max: -Infinity },
-  );
-  const adjustedMarkSizeRatio = Math.min(Math.max(markSizeRatio, 0), 100);
-  const radiusStep = (max - min || max * 100) / Math.pow(adjustedMarkSizeRatio, 2);
-  return function getRadius(mark, defaultRadius = 0): number {
-    if (mark === null) {
-      return defaultRadius;
-    }
-    const circleRadius = (mark / 2 - min) / radiusStep;
-    const baseMagicNumber = 2;
-    return circleRadius ? Math.sqrt(circleRadius + baseMagicNumber) + lineWidth : lineWidth;
-  };
-}
-
-function renderPoints(
-  shift: number,
-  dataSeries: DataSeries,
-  xScale: Scale,
-  yScale: Scale,
-  panel: Dimensions,
-  color: Color,
-  lineStyle: LineStyle,
-  hasY0Accessors: boolean,
-  markSizeOptions: MarkSizeOptions,
-  styleAccessor?: PointStyleAccessor,
-  spatial = false,
-): {
-  pointGeometries: PointGeometry[];
-  indexedGeometryMap: IndexedGeometryMap;
-} {
-  const indexedGeometryMap = new IndexedGeometryMap();
-  const isLogScale = isLogarithmicScale(yScale);
-  const getRadius = markSizeOptions.enabled
-    ? getRadiusFn(dataSeries.data, lineStyle.strokeWidth, markSizeOptions.ratio)
-    : () => 0;
-  const geometryType = spatial ? GeometryType.spatial : GeometryType.linear;
-  const pointGeometries = dataSeries.data.reduce((acc, datum) => {
-    const { x: xValue, y0, y1, mark } = datum;
-    // don't create the point if not within the xScale domain
-    if (!xScale.isValueInDomain(xValue)) {
-      return acc;
-    }
-    // don't create the point if it that point was filled
-    if (isDatumFilled(datum)) {
-      return acc;
-    }
-    const x = xScale.scale(xValue);
-
-    if (x === null) {
-      return acc;
-    }
-
-    const points: PointGeometry[] = [];
-    const yDatums = hasY0Accessors ? [y0, y1] : [y1];
-
-    yDatums.forEach((yDatum, index) => {
-      // skip rendering point if y1 is null
-      if (y1 === null) {
-        return;
-      }
-      let y;
-      let radius = getRadius(mark);
-      // we fix 0 and negative values at y = 0
-      if (yDatum === null || (isLogScale && yDatum <= 0)) {
-        [y] = yScale.range;
-        radius = 0;
-      } else {
-        y = yScale.scale(yDatum);
-      }
-
-      if (y === null) {
-        return acc;
-      }
-      const originalY = getDatumYValue(datum, index === 0, hasY0Accessors, dataSeries.stackMode);
-      const seriesIdentifier: XYChartSeriesIdentifier = {
-        key: dataSeries.key,
-        specId: dataSeries.specId,
-        yAccessor: dataSeries.yAccessor,
-        splitAccessors: dataSeries.splitAccessors,
-        seriesKeys: dataSeries.seriesKeys,
-        smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
-        smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
-      };
-      const styleOverrides = getPointStyleOverrides(datum, seriesIdentifier, styleAccessor);
-      const pointGeometry: PointGeometry = {
-        radius,
-        x,
-        y,
-        color,
-        value: {
-          x: xValue,
-          y: originalY,
-          mark,
-          accessor: hasY0Accessors && index === 0 ? BandedAccessorType.Y0 : BandedAccessorType.Y1,
-          datum: datum.datum,
-        },
-        transform: {
-          x: shift,
-          y: 0,
-        },
-        seriesIdentifier,
-        styleOverrides,
-        panel,
-      };
-      indexedGeometryMap.set(pointGeometry, geometryType);
-      // use the geometry only if the yDatum in contained in the current yScale domain
-      const isHidden = yDatum === null || (isLogScale && yDatum <= 0);
-      if (!isHidden && yScale.isValueInDomain(yDatum)) {
-        points.push(pointGeometry);
-      }
-    });
-    return [...acc, ...points];
-  }, [] as PointGeometry[]);
-  return {
-    pointGeometries,
-    indexedGeometryMap,
-  };
-}
-
-/**
- * Get the original/initial Y value from the datum
- * @param datum a DataSeriesDatum
- * @param lookingForY0 if we are interested in the y0 value, false for y1
- * @param isBandChart if the chart is a band chart
- * @param stackMode an optional stack mode
- */
-function getDatumYValue(
-  { y1, y0, initialY1, initialY0 }: DataSeriesDatum,
-  lookingForY0: boolean,
-  isBandChart: boolean,
-  stackMode?: StackMode,
-) {
-  if (isBandChart) {
-    return stackMode === StackMode.Percentage
-      ? // on band stacked charts in percentage mode, the values I'm looking for are the percentage value
-        // that are already computed and available on y0 and y1
-        lookingForY0
-        ? y0
-        : y1
-      : // in all other cases for band charts, I want to get back the original/initial value of y0 and y1
-      // not the computed value
-      lookingForY0
-      ? initialY0
-      : initialY1;
-  }
-  // if not a band chart get use the original/initial value in every case except for stack as percentage
-  // in this case, we should take the difference between the bottom position of the bar and the top position
-  // of the bar
-  return stackMode === StackMode.Percentage ? (y1 ?? 0) - (y0 ?? 0) : initialY1;
-}
-
-/** @internal */
-export function renderBars(
-  orderIndex: number,
-  dataSeries: DataSeries,
-  xScale: Scale,
-  yScale: Scale,
-  panel: Dimensions,
-  color: Color,
-  sharedSeriesStyle: BarSeriesStyle,
-  displayValueSettings?: DisplayValueSpec,
-  styleAccessor?: BarStyleAccessor,
-  minBarHeight?: number,
-  stackMode?: StackMode,
-  chartRotation?: number,
-): {
-  barGeometries: BarGeometry[];
-  indexedGeometryMap: IndexedGeometryMap;
-} {
-  const indexedGeometryMap = new IndexedGeometryMap();
-  const barGeometries: BarGeometry[] = [];
-
-  const bboxCalculator = new CanvasTextBBoxCalculator();
-
-  // default padding to 1 for now
-  const padding = 1;
-  const { fontSize, fontFamily } = sharedSeriesStyle.displayValue;
-  const absMinHeight = minBarHeight && Math.abs(minBarHeight);
-
-  dataSeries.data.forEach((datum) => {
-    const { y0, y1, initialY1, filled } = datum;
-    // don't create a bar if the initialY1 value is null.
-    if (y1 === null || initialY1 === null || (filled && filled.y1 !== undefined)) {
-      return;
-    }
-    // don't create a bar if not within the xScale domain
-    if (!xScale.isValueInDomain(datum.x)) {
-      return;
-    }
-
-    let y: number | null;
-    let y0Scaled;
-    if (yScale.type === ScaleType.Log) {
-      y = y1 === 0 || y1 === null ? yScale.range[0] : yScale.scale(y1);
-      if (yScale.isInverted) {
-        y0Scaled = y0 === 0 || y0 === null ? yScale.range[1] : yScale.scale(y0);
-      } else {
-        y0Scaled = y0 === 0 || y0 === null ? yScale.range[0] : yScale.scale(y0);
-      }
-    } else {
-      y = yScale.scale(y1);
-      if (yScale.isInverted) {
-        // use always zero as baseline if y0 is null
-        y0Scaled = y0 === null ? yScale.scale(0) : yScale.scale(y0);
-      } else {
-        y0Scaled = y0 === null ? yScale.scale(0) : yScale.scale(y0);
-      }
-    }
-
-    if (y === null || y0Scaled === null) {
-      return;
-    }
-    let height = y0Scaled - y;
-
-    // handle minBarHeight adjustment
-    if (absMinHeight !== undefined && height !== 0 && Math.abs(height) < absMinHeight) {
-      const heightDelta = absMinHeight - Math.abs(height);
-      if (height < 0) {
-        height = -absMinHeight;
-        y += heightDelta;
-      } else {
-        height = absMinHeight;
-        y -= heightDelta;
-      }
-    }
-
-    const xScaled = xScale.scale(datum.x);
-
-    if (xScaled === null) {
-      return;
-    }
-
-    const x = xScaled + xScale.bandwidth * orderIndex;
-    const width = xScale.bandwidth;
-    const originalY1Value = stackMode === StackMode.Percentage ? y1 - (y0 ?? 0) : initialY1;
-    const formattedDisplayValue =
-      displayValueSettings && displayValueSettings.valueFormatter
-        ? displayValueSettings.valueFormatter(originalY1Value)
-        : undefined;
-
-    // only show displayValue for even bars if showOverlappingValue
-    const displayValueText =
-      displayValueSettings && displayValueSettings.isAlternatingValueLabel && barGeometries.length % 2
-        ? undefined
-        : formattedDisplayValue;
-
-    const { displayValueWidth, fixedFontScale } = computeBoxWidth(
-      displayValueText || '',
-      { padding, fontSize, fontFamily, bboxCalculator, width },
-      displayValueSettings,
-    );
-
-    const isHorizontalRotation = chartRotation == null || [0, 180].includes(chartRotation);
-    // Take 70% of space for the label text
-    const fontSizeFactor = 0.7;
-    // Pick the right side of the label's box to use as factor reference
-    const referenceWidth = Math.max(isHorizontalRotation ? displayValueWidth : fixedFontScale, 1);
-
-    const textScalingFactor = getFinalFontScalingFactor(
-      (width * fontSizeFactor) / referenceWidth,
-      fixedFontScale,
-      fontSize,
-    );
-
-    const hideClippedValue = displayValueSettings ? displayValueSettings.hideClippedValue : undefined;
-    // Based on rotation scale the width of the text box
-    const bboxWidthFactor = isHorizontalRotation ? textScalingFactor : 1;
-
-    const displayValue =
-      displayValueSettings && displayValueSettings.showValueLabel
-        ? {
-            fontScale: textScalingFactor,
-            fontSize: fixedFontScale,
-            text: displayValueText,
-            width: bboxWidthFactor * displayValueWidth,
-            height: textScalingFactor * fixedFontScale,
-            hideClippedValue,
-            isValueContainedInElement: displayValueSettings.isValueContainedInElement,
-          }
-        : undefined;
-
-    const seriesIdentifier: XYChartSeriesIdentifier = {
-      key: dataSeries.key,
-      specId: dataSeries.specId,
-      yAccessor: dataSeries.yAccessor,
-      splitAccessors: dataSeries.splitAccessors,
-      seriesKeys: dataSeries.seriesKeys,
-      smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
-      smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
-    };
-
-    const seriesStyle = getBarStyleOverrides(datum, seriesIdentifier, sharedSeriesStyle, styleAccessor);
-
-    const barGeometry: BarGeometry = {
-      displayValue,
-      x,
-      y,
-      transform: {
-        x: 0,
-        y: 0,
-      },
-      width,
-      height,
-      color,
-      value: {
-        x: datum.x,
-        y: originalY1Value,
-        mark: null,
-        accessor: BandedAccessorType.Y1,
-        datum: datum.datum,
-      },
-      seriesIdentifier,
-      seriesStyle,
-      panel,
-    };
-    indexedGeometryMap.set(barGeometry);
-    barGeometries.push(barGeometry);
-  });
-
-  bboxCalculator.destroy();
-
-  return {
-    barGeometries,
-    indexedGeometryMap,
-  };
-}
-
-/** @internal */
-export function renderLine(
-  shift: number,
-  dataSeries: DataSeries,
-  xScale: Scale,
-  yScale: Scale,
-  panel: Dimensions,
-  color: Color,
-  curve: CurveType,
-  hasY0Accessors: boolean,
-  xScaleOffset: number,
-  seriesStyle: LineSeriesStyle,
-  markSizeOptions: MarkSizeOptions,
-  pointStyleAccessor?: PointStyleAccessor,
-  hasFit?: boolean,
-): {
-  lineGeometry: LineGeometry;
-  indexedGeometryMap: IndexedGeometryMap;
-} {
-  const isLogScale = isLogarithmicScale(yScale);
-  const pathGenerator = line<DataSeriesDatum>()
-    .x(({ x }) => xScale.scaleOrThrow(x) - xScaleOffset)
-    .y((datum) => {
-      const yValue = getYValue(datum);
-
-      if (yValue !== null) {
-        return yScale.scaleOrThrow(yValue);
-      }
-
-      // this should never happen thanks to the defined function
-      return yScale.isInverted ? yScale.range[1] : yScale.range[0];
-    })
-    .defined((datum) => {
-      const yValue = getYValue(datum);
-      return yValue !== null && !(isLogScale && yValue <= 0) && xScale.isValueInDomain(datum.x);
-    })
-    .curve(getCurveFactory(curve));
-
-  const { pointGeometries, indexedGeometryMap } = renderPoints(
-    shift - xScaleOffset,
-    dataSeries,
-    xScale,
-    yScale,
-    panel,
-    color,
-    seriesStyle.line,
-    hasY0Accessors,
-    markSizeOptions,
-    pointStyleAccessor,
-  );
-
-  const clippedRanges = getClippedRanges(dataSeries.data, xScale, xScaleOffset);
-  let linePath: string;
-
-  try {
-    linePath = pathGenerator(dataSeries.data) || '';
-  } catch {
-    // When values are not scalable
-    linePath = '';
-  }
-
-  const lineGeometry = {
-    line: linePath,
-    points: pointGeometries,
-    color,
-    transform: {
-      x: shift,
-      y: 0,
-    },
-    seriesIdentifier: {
-      key: dataSeries.key,
-      specId: dataSeries.specId,
-      yAccessor: dataSeries.yAccessor,
-      splitAccessors: dataSeries.splitAccessors,
-      seriesKeys: dataSeries.seriesKeys,
-      smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
-      smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
-    },
-    seriesLineStyle: seriesStyle.line,
-    seriesPointStyle: seriesStyle.point,
-    clippedRanges,
-    hideClippedRanges: !hasFit,
-  };
-  return {
-    lineGeometry,
-    indexedGeometryMap,
-  };
-}
-
-/** @internal */
-export function renderBubble(
-  shift: number,
-  dataSeries: DataSeries,
-  xScale: Scale,
-  yScale: Scale,
-  color: Color,
-  panel: Dimensions,
-  hasY0Accessors: boolean,
-  xScaleOffset: number,
-  seriesStyle: BubbleSeriesStyle,
-  markSizeOptions: MarkSizeOptions,
-  isMixedChart: boolean,
-  pointStyleAccessor?: PointStyleAccessor,
-): {
-  bubbleGeometry: BubbleGeometry;
-  indexedGeometryMap: IndexedGeometryMap;
-} {
-  const { pointGeometries, indexedGeometryMap } = renderPoints(
-    shift - xScaleOffset,
-    dataSeries,
-    xScale,
-    yScale,
-    panel,
-    color,
-    seriesStyle.point,
-    hasY0Accessors,
-    markSizeOptions,
-    pointStyleAccessor,
-    !isMixedChart,
-  );
-
-  const bubbleGeometry = {
-    points: pointGeometries,
-    color,
-    seriesIdentifier: {
-      key: dataSeries.key,
-      specId: dataSeries.specId,
-      yAccessor: dataSeries.yAccessor,
-      splitAccessors: dataSeries.splitAccessors,
-      seriesKeys: dataSeries.seriesKeys,
-      smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
-      smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
-    },
-    seriesPointStyle: seriesStyle.point,
-  };
-  return {
-    bubbleGeometry,
-    indexedGeometryMap,
-  };
-}
-
-/** @internal */
-
-export function renderArea(
-  shift: number,
-  dataSeries: DataSeries,
-  xScale: Scale,
-  yScale: Scale,
-  panel: Dimensions,
-  color: Color,
-  curve: CurveType,
-  hasY0Accessors: boolean,
-  xScaleOffset: number,
-  seriesStyle: AreaSeriesStyle,
-  markSizeOptions: MarkSizeOptions,
-  isStacked = false,
-  pointStyleAccessor?: PointStyleAccessor,
-  hasFit?: boolean,
-): {
-  areaGeometry: AreaGeometry;
-  indexedGeometryMap: IndexedGeometryMap;
-} {
-  const isLogScale = isLogarithmicScale(yScale);
-
-  const pathGenerator = area<DataSeriesDatum>()
-    .x(({ x }) => xScale.scaleOrThrow(x) - xScaleOffset)
-    .y1((datum) => {
-      const yValue = getYValue(datum);
-      if (yValue !== null) {
-        return yScale.scaleOrThrow(yValue);
-      }
-      // this should never happen thanks to the defined function
-      return yScale.isInverted ? yScale.range[1] : yScale.range[0];
-    })
-    .y0(({ y0 }) => {
-      if (y0 === null) {
-        if (isLogScale) {
-          return yScale.scaleOrThrow(1);
-        }
-        return yScale.scaleOrThrow(0);
-      }
-      if (isLogScale) {
-        return yScale.scaleOrThrow(Math.max(1, y0));
-      }
-      return yScale.scaleOrThrow(y0);
-    })
-    .defined((datum) => {
-      const yValue = getYValue(datum);
-      return yValue !== null && !(isLogScale && yValue <= 0) && xScale.isValueInDomain(datum.x);
-    })
-    .curve(getCurveFactory(curve));
-
-  const clippedRanges = getClippedRanges(dataSeries.data, xScale, xScaleOffset);
-
-  let y1Line: string | null;
-
-  try {
-    y1Line = pathGenerator.lineY1()(dataSeries.data);
-  } catch {
-    // When values are not scalable
-    y1Line = null;
-  }
-
-  const lines: string[] = [];
-  if (y1Line) {
-    lines.push(y1Line);
-  }
-  if (hasY0Accessors) {
-    let y0Line: string | null;
-
-    try {
-      y0Line = pathGenerator.lineY0()(dataSeries.data);
-    } catch {
-      // When values are not scalable
-      y0Line = null;
-    }
-    if (y0Line) {
-      lines.push(y0Line);
-    }
-  }
-
-  const { pointGeometries, indexedGeometryMap } = renderPoints(
-    shift - xScaleOffset,
-    dataSeries,
-    xScale,
-    yScale,
-    panel,
-    color,
-    seriesStyle.line,
-    hasY0Accessors,
-    markSizeOptions,
-    pointStyleAccessor,
-    false,
-  );
-
-  let areaPath: string;
-
-  try {
-    areaPath = pathGenerator(dataSeries.data) || '';
-  } catch {
-    // When values are not scalable
-    areaPath = '';
-  }
-
-  const areaGeometry: AreaGeometry = {
-    area: areaPath,
-    lines,
-    points: pointGeometries,
-    color,
-    transform: {
-      y: 0,
-      x: shift,
-    },
-    seriesIdentifier: {
-      key: dataSeries.key,
-      specId: dataSeries.specId,
-      yAccessor: dataSeries.yAccessor,
-      splitAccessors: dataSeries.splitAccessors,
-      seriesKeys: dataSeries.seriesKeys,
-      smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
-      smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
-    },
-    seriesAreaStyle: seriesStyle.area,
-    seriesAreaLineStyle: seriesStyle.line,
-    seriesPointStyle: seriesStyle.point,
-    isStacked,
-    clippedRanges,
-    hideClippedRanges: !hasFit,
-  };
-  return {
-    areaGeometry,
-    indexedGeometryMap,
-  };
-}
-
-/**
- *
- * @param param0
- * @internal
- */
-export function isDatumFilled({ filled, initialY1 }: DataSeriesDatum) {
-  return filled?.x !== undefined || filled?.y1 !== undefined || initialY1 === null || initialY1 === undefined;
-}
-
-/**
- * Gets clipped ranges that have been fitted to values
- * @param dataset
- * @param xScale
- * @param xScaleOffset
- * @internal
- */
-export function getClippedRanges(dataset: DataSeriesDatum[], xScale: Scale, xScaleOffset: number): ClippedRanges {
-  let firstNonNullX: number | null = null;
-  let hasNull = false;
-  return dataset.reduce<ClippedRanges>((acc, data) => {
-    const xScaled = xScale.scale(data.x);
-    if (xScaled === null) {
-      return acc;
-    }
-
-    const xValue = xScaled - xScaleOffset + xScale.bandwidth / 2;
-
-    if (isDatumFilled(data)) {
-      const endXValue = xScale.range[1] - xScale.bandwidth * (2 / 3);
-      if (firstNonNullX !== null && xValue === endXValue) {
-        acc.push([firstNonNullX, xValue]);
-      }
-      hasNull = true;
-    } else {
-      if (hasNull) {
-        if (firstNonNullX !== null) {
-          acc.push([firstNonNullX, xValue]);
-        } else {
-          acc.push([0, xValue]);
-        }
-        hasNull = false;
-      }
-
-      firstNonNullX = xValue;
-    }
-    return acc;
-  }, []);
-}
-
-/** @internal */
-export function getGeometryStateStyle(
-  seriesIdentifier: XYChartSeriesIdentifier,
-  sharedGeometryStyle: SharedGeometryStateStyle,
-  highlightedLegendItem?: LegendItem,
-  individualHighlight?: { [key: string]: boolean },
-): GeometryStateStyle {
-  const { default: defaultStyles, highlighted, unhighlighted } = sharedGeometryStyle;
-
-  if (highlightedLegendItem) {
-    const isPartOfHighlightedSeries = seriesIdentifier.key === highlightedLegendItem.seriesIdentifier.key;
-
-    return isPartOfHighlightedSeries ? highlighted : unhighlighted;
-  }
-
-  if (individualHighlight) {
-    const { hasHighlight, hasGeometryHover } = individualHighlight;
-    if (!hasGeometryHover) {
-      return highlighted;
-    }
-    return hasHighlight ? highlighted : unhighlighted;
-  }
-
-  return defaultStyles;
-}
-
-/** @internal */
-export function isPointOnGeometry(
-  xCoordinate: number,
-  yCoordinate: number,
-  indexedGeometry: BarGeometry | PointGeometry,
-  buffer: MarkBuffer = DEFAULT_HIGHLIGHT_PADDING,
-) {
-  const { x, y } = indexedGeometry;
-  if (isPointGeometry(indexedGeometry)) {
-    const { radius } = indexedGeometry;
-    const distance = getDistance(
-      {
-        x: xCoordinate,
-        y: yCoordinate,
-      },
-      {
-        x,
-        y,
-      },
-    );
-
-    const radiusBuffer = typeof buffer === 'number' ? buffer : buffer(radius);
-
-    if (radiusBuffer === Infinity) {
-      return distance <= radius + DEFAULT_HIGHLIGHT_PADDING;
-    }
-
-    return distance <= radius + radiusBuffer;
-  }
-  const { width, height } = indexedGeometry;
-  return yCoordinate >= y && yCoordinate <= y + height && xCoordinate >= x && xCoordinate <= x + width;
-}
