+++ src/chart_types/xy_chart/rendering/utils.ts
@@ -1,0 +1,156 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { LegendItem } from '../../../commons/legend';
+import { Scale } from '../../../scales';
+import { MarkBuffer } from '../../../specs';
+import { getDistance } from '../../../utils/commons';
+import { BarGeometry, ClippedRanges, isPointGeometry, PointGeometry } from '../../../utils/geometry';
+import { GeometryStateStyle, SharedGeometryStateStyle } from '../../../utils/themes/theme';
+import { DataSeriesDatum, XYChartSeriesIdentifier } from '../utils/series';
+import { DEFAULT_HIGHLIGHT_PADDING } from './constants';
+
+export interface MarkSizeOptions {
+  enabled: boolean;
+  ratio?: number;
+}
+
+/**
+ * Returns value of `y1` or `filled.y1` or null
+ * @internal
+ */
+export const getYValue = ({ y1, filled }: DataSeriesDatum): number | null => {
+  if (y1 !== null) {
+    return y1;
+  }
+
+  if (filled && filled.y1 !== undefined) {
+    return filled.y1;
+  }
+
+  return null;
+};
+
+/**
+ *
+ * @param param0
+ * @internal
+ */
+export function isDatumFilled({ filled, initialY1 }: DataSeriesDatum) {
+  return filled?.x !== undefined || filled?.y1 !== undefined || initialY1 === null || initialY1 === undefined;
+}
+
+/**
+ * Gets clipped ranges that have been fitted to values
+ * @param dataset
+ * @param xScale
+ * @param xScaleOffset
+ * @internal
+ */
+export function getClippedRanges(dataset: DataSeriesDatum[], xScale: Scale, xScaleOffset: number): ClippedRanges {
+  let firstNonNullX: number | null = null;
+  let hasNull = false;
+  return dataset.reduce<ClippedRanges>((acc, data) => {
+    const xScaled = xScale.scale(data.x);
+    if (xScaled === null) {
+      return acc;
+    }
+
+    const xValue = xScaled - xScaleOffset + xScale.bandwidth / 2;
+
+    if (isDatumFilled(data)) {
+      const endXValue = xScale.range[1] - xScale.bandwidth * (2 / 3);
+      if (firstNonNullX !== null && xValue === endXValue) {
+        acc.push([firstNonNullX, xValue]);
+      }
+      hasNull = true;
+    } else {
+      if (hasNull) {
+        if (firstNonNullX !== null) {
+          acc.push([firstNonNullX, xValue]);
+        } else {
+          acc.push([0, xValue]);
+        }
+        hasNull = false;
+      }
+
+      firstNonNullX = xValue;
+    }
+    return acc;
+  }, []);
+}
+
+/** @internal */
+export function getGeometryStateStyle(
+  seriesIdentifier: XYChartSeriesIdentifier,
+  sharedGeometryStyle: SharedGeometryStateStyle,
+  highlightedLegendItem?: LegendItem,
+  individualHighlight?: { [key: string]: boolean },
+): GeometryStateStyle {
+  const { default: defaultStyles, highlighted, unhighlighted } = sharedGeometryStyle;
+
+  if (highlightedLegendItem) {
+    const isPartOfHighlightedSeries = seriesIdentifier.key === highlightedLegendItem.seriesIdentifier.key;
+
+    return isPartOfHighlightedSeries ? highlighted : unhighlighted;
+  }
+
+  if (individualHighlight) {
+    const { hasHighlight, hasGeometryHover } = individualHighlight;
+    if (!hasGeometryHover) {
+      return highlighted;
+    }
+    return hasHighlight ? highlighted : unhighlighted;
+  }
+
+  return defaultStyles;
+}
+
+/** @internal */
+export function isPointOnGeometry(
+  xCoordinate: number,
+  yCoordinate: number,
+  indexedGeometry: BarGeometry | PointGeometry,
+  buffer: MarkBuffer = DEFAULT_HIGHLIGHT_PADDING,
+) {
+  const { x, y } = indexedGeometry;
+  if (isPointGeometry(indexedGeometry)) {
+    const { radius } = indexedGeometry;
+    const distance = getDistance(
+      {
+        x: xCoordinate,
+        y: yCoordinate,
+      },
+      {
+        x,
+        y,
+      },
+    );
+
+    const radiusBuffer = typeof buffer === 'number' ? buffer : buffer(radius);
+
+    if (radiusBuffer === Infinity) {
+      return distance <= radius + DEFAULT_HIGHLIGHT_PADDING;
+    }
+
+    return distance <= radius + radiusBuffer;
+  }
+  const { width, height } = indexedGeometry;
+  return yCoordinate >= y && yCoordinate <= y + height && xCoordinate >= x && xCoordinate <= x + width;
+}
