+++ src/chart_types/xy_chart/state/utils/utils.ts
@@ -39,7 +39,11 @@ import { getPredicateFn, Predicate } from '../../../heatmap/utils/commons';
 import { XDomain } from '../../domains/types';
 import { mergeXDomain } from '../../domains/x_domain';
 import { isStackedSpec, mergeYDomain } from '../../domains/y_domain';
-import { renderArea, renderBars, renderLine, renderBubble, isDatumFilled } from '../../rendering/rendering';
+import { renderArea } from '../../rendering/area';
+import { renderBars } from '../../rendering/bars';
+import { renderBubble } from '../../rendering/bubble';
+import { renderLine } from '../../rendering/line';
+import { isDatumFilled } from '../../rendering/utils';
 import { defaultTickFormatter } from '../../utils/axis_utils';
 import { fillSeries } from '../../utils/fill_series';
 import { groupBy } from '../../utils/group_data_series';
